+++ lib/mongoid/history/tracker.rb
@@ -14,7 +14,11 @@ module Mongoid
         field :version,                 type: Integer
         field :action,                  type: String
         field :scope,                   type: String
-        belongs_to :modifier, class_name: Mongoid::History.modifier_class_name
+        modifier_options = {
+          class_name: Mongoid::History.modifier_class_name
+        }
+        modifier_options[:optional] = true if Mongoid::Compatibility::Version.mongoid6_or_newer?
+        belongs_to :modifier, modifier_options
 
         index(scope: 1)
         index(association_chain: 1)
@@ -50,7 +54,7 @@ module Mongoid
         undo_hash = affected.easy_unmerge(modified)
         undo_hash.easy_merge!(original)
         modifier_field = trackable.history_trackable_options[:modifier_field]
-        undo_hash[modifier_field] = modifier
+        undo_hash[modifier_field] = modifier if modifier_field
         (modified.keys - undo_hash.keys).each do |k|
           undo_hash[k] = nil
         end
@@ -61,7 +65,7 @@ module Mongoid
         redo_hash = affected.easy_unmerge(original)
         redo_hash.easy_merge!(modified)
         modifier_field = trackable.history_trackable_options[:modifier_field]
-        redo_hash[modifier_field] = modifier
+        redo_hash[modifier_field] = modifier if modifier_field
         localize_keys(redo_hash)
       end